+ Implemento ModeloFactory::rutinaAgregarNuevoCliente().- + Implemento ModeloFactory::crearJugador().- + Implemento HiloConfiguracion::rutina().- + Habilito lectura de varios protagonistas desde el ParserYaml.- + Paso enviarArchivosDeConfiguracion() y enviarListaDeArchivos() de HiloConfiguracion a ModeloFactory.- + Cambios menores en las demás clases.-

+++ TP2/src/utils/Sockets/ConexionCliente.cpp
@@ -42,7 +42,7 @@ void ConexionCliente::iniciarEnvio(void){
 	this->hiloSalida.correrConexion(iniciarParametrosRun('S',NULL,NULL,NULL));
 }
 
-long ConexionCliente::getId(void){
+int ConexionCliente::getId(void){
 	return this->id;
 }
